requirements.txt
Flask==2.3.3
opencv-python==4.10.0.84
numpy==1.24.3
Pillow==10.0.1
pandas==2.1.1
openpyxl==3.1.2
python-dateutil==2.8.2
Werkzeug==2.3.7
face-recognition==1.3.0
dlib==19.24.2
scikit-learn==1.3.0
pickle-mixin==1.0.2
